archive/pyproject.toml
[project]
name = "tap-db2"
dependencies = [
        "attrs==23.1.0",
        "ibm-db-sa==0.4.0",
        "ibm-db==3.1.4",
        "jinja2==3.1.2",
        "markupsafe<2.2.0",
        "pendulum==2.1.2",
        "pyodbc==4.0.39",
        "pytz>=2018.1",
        "singer-python>=5.12.0",
        "sqlalchemy<3.0.0",
]
dynamic = [
   "version"
]

[build-system]
requires = ["setuptools", "wheel", "setuptools_scm[toml]>=6.2"]

[project.scripts]
tap-db2 = "tap_db2:main"

[tool.setuptools_scm]
